requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
PyJWT==2.8.0
google-auth==2.22.0
requests==2.31.0